Setup authentication. Refactor Tokenization to be inside Token
