+++ middleware/authenticate.js
@@ -2,21 +2,13 @@ const jwt = require('jsonwebtoken')
 const { ObjectID } = require('mongodb')
 const mongoose = require('mongoose')
 
-var authenticateToken = (req, res, next) => {
+let authenticate = (req, res, next) => {
   let rawToken = req.header('x-auth')
 
   let unvalidatedHeader = jwt.decode(rawToken)
 
   let unvalidatedTokenId = unvalidatedHeader._id
   let unvalidatedUserType = unvalidatedHeader.access
-
-  if (unvalidatedUserType === 'Student' && req.userType === '') { // allows instructors access to student sections
-    res.status(401).send('Mismatched credentials')
-  }
-
-  if (!ObjectID(unvalidatedTokenId)) {
-    res.status(401).send('Invalid token id') // send access denied
-  }
 
   mongoose.model(unvalidatedUserType).findOne({
     'token._id': unvalidatedTokenId
@@ -37,8 +29,8 @@ var authenticateToken = (req, res, next) => {
     } catch (error) {
       return Promise.reject(new TypeError('Token validation failed'))
     }
-  }).catch((e) => {
-    res.status(401).send('Invalid token')
+  }).catch(() => {
+    res.status(404).send('invalid token')
   })
 }
 
@@ -46,14 +38,14 @@ let authenticateStudent = (req, res, next) => {
   console.log('Authenticating Student')
   req.userType = 'Student'
 
-  authenticateToken(req, res, next)
+  authenticate(req, res, next)
 }
 
 let authenticateInstructor = (req, res, next) => {
   console.log('Authenticating Instructor')
   req.userType = 'Instructor'
 
-  authenticateInstructor(req, res, next)
+  authenticate(req, res, next)
 }
 
 module.exports = { authenticateStudent, authenticateInstructor }
