+++ models/student.js
@@ -1,7 +1,5 @@
 const mongoose = require('mongoose')
-const { Token, TokenSchema } = require('./token')
-const jwt = require('jsonwebtoken')
-const bcrypt = require('bcrypt')
+const { TokenSchema } = require('./token')
 
 let StudentSchema = new mongoose.Schema({
   classcode: {
@@ -18,36 +16,6 @@ let StudentSchema = new mongoose.Schema({
   },
   token: TokenSchema
 })
-
-StudentSchema.methods.generateTokenAndSave = function (callback) {
-  console.log('generating token and saving')
-
-  let student = this
-  let access = 'Student'
-
-  bcrypt.genSalt(12, (err, salt) => {
-    if (err) {
-      return Promise.reject(new TypeError())
-    }
-
-    let tokenId = mongoose.Types.ObjectId()
-
-    let tokenVal = jwt.sign({ _id: tokenId, access }, salt).toString()
-
-    let token = new Token({
-      _id: tokenId,
-      token: tokenVal,
-      access,
-      salt
-    })
-
-    student.token = token
-
-    student.save()
-
-    callback()
-  })
-}
 
 let Student = mongoose.model('Student', StudentSchema)
