+++ models/token.js
@@ -1,4 +1,6 @@
 const mongoose = require('mongoose')
+const bcrypt = require('bcrypt')
+const jwt = require('jsonwebtoken')
 
 let TokenSchema = new mongoose.Schema({
   _id: {
@@ -19,6 +21,33 @@ let TokenSchema = new mongoose.Schema({
     required: true
   }
 })
+
+TokenSchema.statics.generateAuthToken = function (access) {
+  console.log(access)
+
+  let Token = this  
+
+  return new Promise((resolve, reject) => {
+    bcrypt.genSalt(12, (err, salt) => {
+      if (err) {
+        reject(err)
+      }
+
+      let tokenId = mongoose.Types.ObjectId()
+
+      let tokenVal = jwt.sign({ _id: tokenId, access }, salt, { expiresIn: '4h' }).toString()
+
+      let token = new Token({
+        _id: tokenId,
+        token: tokenVal,
+        access,
+        salt
+      })
+
+      resolve(token)
+    })
+  })
+}
 
 let Token = mongoose.model('Token', TokenSchema)
